Always rely on SD when searching for a service #34006
Have a server register a service then unregister it.
It happens a race condition if a client is trying to get the unregister
service just after it is unregister.
SessionService has a local remote object on all services request and clean
this local remote object on signal connect to the SD.

Change-Id: I3cb0acfd371cd0a102e37476d4ed7042bfc40f66